% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_plot.R
\name{basic_plot}
\alias{basic_plot}
\title{Create a scatterplot to view your data using geom_jitter to compare variances.}
\usage{
basic_plot(data, x, y)
}
\arguments{
\item{data}{The dataset you read in (dataframe)}

\item{x}{The independent variable}

\item{y}{The dependent variable}
}
\value{
myplot Scatter plot
}
\description{
Use a scatterplot to view your data before performing analyses. Will also save a copy of the plot.
}
