% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_model}
\alias{linear_model}
\title{Create a linear model regression summary for your data.}
\usage{
linear_model(data, dependent, independent_cols)
}
\arguments{
\item{data}{The dataset you read in (dataframe)}

\item{dependent}{dependent column}

\item{independent_cols}{independent columns}
}
\value{
summary The summary of your linear model regression.
}
\description{
Determine if your variables have a relationship. Summary will provide statistical values for residuals, estimates, standard error, t-value and p-values.
}
