% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_abundance.R
\name{mapping_abundance}
\alias{mapping_abundance}
\title{Mapping abundance function}
\usage{
mapping_abundance(data, Species, Lat, Long, Sector)
}
\arguments{
\item{data}{Dataframe you cleaned}

\item{Species}{The species or common name column}

\item{Lat}{The column with Lat coordinates}

\item{Long}{The column with Long coordinates}

\item{Sector}{From Sectors 1-6}
}
\value{
abundanceMap The interactive map of abundances
}
\description{
Plot abundances in a map. Need the following packages: tidyverse, sf, mapview, and vegan
}
