% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_plot.R
\name{nmds_plot}
\alias{nmds_plot}
\title{Non-metric Multidimensional Scale plot}
\usage{
nmds_plot(data, Species, Sector)
}
\arguments{
\item{data}{The data you cleaned (dataframe)}

\item{Species}{Your species name column}

\item{Sector}{Your sector column}
}
\value{
plot Your nmds plot using ggplot
}
\description{
Creates a ggplot based on non-metric multidimensional scale to show similarity between groups.
}
