% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_abundance.R
\name{rank_abundance}
\alias{rank_abundance}
\title{Show rank abundance by plotting your data in descending order by phylum}
\usage{
rank_abundance(data, x, xlabel)
}
\arguments{
\item{data}{The dataset you read in (dataframe)}

\item{x}{The rank of organisms you want to count. (Species column)}

\item{xlabel}{Name of x-axis label, should be the same as x. Add quotations (String)}
}
\value{
rank_plot The rank abundance plot (bar plot)
}
\description{
Use your data that you read in and the phylogenetic rank you want to count.
}
