% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nas.R
\name{remove_nas}
\alias{remove_nas}
\title{Clean your data by removing NAs}
\usage{
remove_nas(data)
}
\arguments{
\item{data}{The data set you read in to remove NAs (dataframe)}
}
\value{
clean The "cleaned" dataset without NAs (dataframe)
}
\description{
Use your data that you read in to clean data
Returns your data without NAs
Save as a variable to be able to use this clean data
}
